security rules para que un user no pueda acerder a los contactos de otro user

+++ contacts-app/delete.php
@@ -13,7 +13,7 @@ if (!isset($_SESSION["user"])) {
 $id = $_GET["id"];
 
 //selecionando todo desde la base de datos y borra ese id
-$stetament = $conn->prepare("SELECT * FROM contacts WHERE id = :id");
+$stetament = $conn->prepare("SELECT * FROM contacts WHERE id = :id LIMIT 1");
 $stetament->execute([":id" => $id]);
 
 // validacion para los errores
@@ -23,14 +23,14 @@ if ($stetament->rowCount() == 0) {
     return;
 }
 
+// aqui especificamos que un usuario pueda hacerder a sus contactos y no  a los demas 
+$contact = $stetament->fetch(PDO::FETCH_ASSOC);
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+    http_response_code(403);
+    echo ("HTTP 403 NO AUTHORIZADO");
+    return;
+}
+
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
 
-// validaciones para que no te pasen una sql injection
-// $stetament->bindParam(":id", $id);
-// $stetament->execute();
-
-// ejecutando el statament y le pasamos un arry asociativo con un shorcut y nos evitamos bindParam de arriba
-// $stetament->execute([":id" => $id]);
-
-// haciendo la redirecion
 header("Location: home.php");
